requirements.txt
datasets==2.20.0
pandas  ==2.2.2
pyspellchecker==0.8.1

regex==2024.5.15
safetensors==0.4.3 
tokenizers==0.19.1 
transformers==4.41.2

MarkupSafe==2.1.5 
jinja2==3.1.4 
mpmath==1.3.0 
networkx==3.3 
sympy==1.12.1 
torch==2.3.1


evaluate==0.4.2


accelerator==2024.6.11 
bottle==0.12.25 
setproctitle==1.3.3
waitress==3.0.0


accelerate==0.31.0
git+https://github.com/huggingface/accelerate